TST: update tests to reflec all changes
